// test/integration/test/ports/engines/writes/nested_mutations/already_converted/_fixture/one_to_one_optional/parent_id_simple__child_id_simple__child_references_parent_id/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Child: {
      readonly c: CodecTypes['pg/text@1']['output'];
      readonly c_1: CodecTypes['pg/text@1']['output'];
      readonly c_2: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly non_unique: CodecTypes['pg/text@1']['output'] | null;
      readonly parentId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Parent: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly non_unique: CodecTypes['pg/text@1']['output'] | null;
      readonly p: CodecTypes['pg/text@1']['output'];
      readonly p_1: CodecTypes['pg/text@1']['output'];
      readonly p_2: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Child: {
      readonly c: CodecTypes['pg/text@1']['input'];
      readonly c_1: CodecTypes['pg/text@1']['input'];
      readonly c_2: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly non_unique: CodecTypes['pg/text@1']['input'] | null;
      readonly parentId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Parent: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly non_unique: CodecTypes['pg/text@1']['input'] | null;
      readonly p: CodecTypes['pg/text@1']['input'];
      readonly p_1: CodecTypes['pg/text@1']['input'];
      readonly p_2: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Child: {
      readonly c: CodecTypes['pg/text@1']['output'];
      readonly c_1: CodecTypes['pg/text@1']['output'];
      readonly c_2: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly non_unique: CodecTypes['pg/text@1']['output'] | null;
      readonly parentId: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Parent: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly non_unique: CodecTypes['pg/text@1']['output'] | null;
      readonly p: CodecTypes['pg/text@1']['output'];
      readonly p_1: CodecTypes['pg/text@1']['output'];
      readonly p_2: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Child: {
      readonly c: CodecTypes['pg/text@1']['input'];
      readonly c_1: CodecTypes['pg/text@1']['input'];
      readonly c_2: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly non_unique: CodecTypes['pg/text@1']['input'] | null;
      readonly parentId: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Parent: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly non_unique: CodecTypes['pg/text@1']['input'] | null;
      readonly p: CodecTypes['pg/text@1']['input'];
      readonly p_1: CodecTypes['pg/text@1']['input'];
      readonly p_2: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;